fyniq_api/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# WebSocket and Channels
channels==4.0.0
channels-redis==4.1.0
daphne==4.2.1

# Celery and Redis
celery==5.3.4
redis==5.0.1

# API Documentation
drf-yasg==1.21.10

# Cryptocurrency and Trading
ccxt==4.1.77
bit==0.8.0
bitcoin==1.1.42
coincurve==21.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
ta==0.11.0

# HTTP and Networking
aiohttp==3.12.13
requests==2.31.0

# Date and Time
python-dateutil==2.8.2
dateparser==1.2.2

# Security and Cryptography
cryptography==41.0.7
PyJWT==2.8.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==37.4.0

# Monitoring and Logging
psutil==5.9.6

# Development and Debugging
django-debug-toolbar==4.2.0

# Environment and Configuration
python-decouple==3.8

# Additional Utilities
Pillow==10.1.0
joblib==1.5.1

# Enhanced crypto dependencies from fluxor_backend
web3==7.12.1
eth-account>=0.13.6
# bitcoinlib==0.6.0  # Commented out due to numpy version conflict
coinbase-commerce==1.0.0
pycoingecko==3.1.0
hdwallet==2.2.1
bip32utils==0.3.post4
mnemonic==0.20
